Separated deciders from QoSObserver. Added Observer config.
