+++ src/observers/qos_correction.hpp
@@ -5,6 +5,8 @@
 #include <string>
 #include <vector>
 
+#include "filters/executor_age.hpp"
+
 #include "glog/logging.h"
 
 #include "mesos/mesos.hpp"
@@ -14,48 +16,30 @@
 
 #include "messages/serenity.hpp"
 
+#include "serenity/config.hpp"
 #include "serenity/serenity.hpp"
 
-#include "filters/executor_age.hpp"
+#include "observers/deciders/base.hpp"
+#include "observers/deciders/seniority_based.hpp"
 
 namespace mesos {
 namespace serenity {
 
-using ContentionDeciderFunction = Try<QoSCorrections>
-    (ExecutorAgeFilter* ageFilter,
-     const Contentions& currentContentions,
-     const ResourceUsage& currentUsage);
-
-
-/**
- * Convenient base class for contention interpretations.
- * It converts contentions & usage to QoSCorrections.
- *
- * Convenient for debugging and testing different algorithms.
- */
-class ContentionDecider {
+class QoSCorrectionObserverConfig : public SerenityConfig {
  public:
-  virtual ContentionDeciderFunction decide = 0;
-};
+  QoSCorrectionObserverConfig() {
+    this->initDefaults();
+  }
 
+  explicit QoSCorrectionObserverConfig(const SerenityConfig& customCfg) {
+    this->initDefaults();
+    this->applyConfig(customCfg);
+  }
 
-/**
- * Kills all BE executors given in usage.
- */
-class KillAllDecider : public ContentionDecider {
- public:
-  ContentionDeciderFunction decide;
-};
-
-
-/**
- * Checks contentions and choose executors to kill.
- * Currently it calculates mean contention and based on that estimates how
- * many executors we should kill. Executors are sorted by age.
- */
-class SeverityBasedSeniorityDecider : public ContentionDecider {
- public:
-  ContentionDeciderFunction decide;
+  void initDefaults() {
+    this->fields[qos_observer::CONTENTION_COOLDOWN] =
+      qos_observer::DEFAULT_CONTENTION_COOLDOWN;
+  }
 };
 
 
@@ -74,13 +58,15 @@ class QoSCorrectionObserver : public SyncConsumer<Contentions>,
       uint64_t _contentionProducents,
       ExecutorAgeFilter* _ageFilter = new ExecutorAgeFilter(),
       ContentionDecider* _contentionDecider =
-        new SeverityBasedSeniorityDecider())
+        new SeverityBasedSeniorityDecider(),
+      const SerenityConfig& _config = QoSCorrectionObserverConfig())
     : SyncConsumer<Contentions>(_contentionProducents),
       Producer<QoSCorrections>(_consumer),
       currentContentions(None()),
       currentUsage(None()),
       contentionDecider(_contentionDecider),
-      ageFilter(_ageFilter) {}
+      ageFilter(_ageFilter),
+      config(QoSCorrectionObserverConfig(_config)) {}
 
   ~QoSCorrectionObserver();
 
@@ -113,7 +99,7 @@ class QoSCorrectionObserver : public SyncConsumer<Contentions>,
     return (firstAllocation.cpus().get() > secondAllocation.cpus().get());
   }
 
-  static constexpr const char* name = "[SerenityQoS] CorrectionObserver: ";
+  static constexpr const char* NAME = "CorrectionObserver: ";
 
  protected:
   Option<Contentions> currentContentions;
@@ -121,38 +107,11 @@ class QoSCorrectionObserver : public SyncConsumer<Contentions>,
   ContentionDecider* contentionDecider;
   ExecutorAgeFilter* ageFilter;
 
+  const SerenityConfig config;
+
   //! Run when all required info are gathered.
   Try<Nothing> __correctSlave();
 };
-
-
-inline std::list<ResourceUsage_Executor> filterPrExecutors(
-    ResourceUsage usage) {
-  std::list<ResourceUsage_Executor> beExecutors;
-  for (ResourceUsage_Executor inExec : usage.executors()) {
-    if (!inExec.has_executor_info()) {
-      LOG(ERROR) << QoSCorrectionObserver::name << "Executor <unknown>"
-      << " does not include executor_info";
-      // Filter out these executors.
-      continue;
-    }
-    if (inExec.allocated().size() == 0) {
-      LOG(ERROR) << QoSCorrectionObserver::name << "Executor "
-      << inExec.executor_info().executor_id().value()
-      << " does not include allocated resources.";
-      // Filter out these executors.
-      continue;
-    }
-
-    Resources allocated(inExec.allocated());
-    // Check if task uses revocable resources.
-    if (!allocated.revocable().empty())
-      beExecutors.push_back(inExec);
-  }
-
-  return beExecutors;
-}
-
 
 }  // namespace serenity
 }  // namespace mesos
